refactor: separate Card create page slice
