+++ src/features/card-form/ui/CardFields.stories.tsx
@@ -1,25 +1,22 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { useEffect } from "react";
 import { useForm } from "react-hook-form";
-import { expect, fn } from "storybook/test";
+import { expect } from "storybook/test";
 
 import type { Card } from "@/entities/card";
 import { CATEGORY } from "@/entities/deck";
 import { withPageLayout } from "@/storybook/PageLayoutDecorator";
 import * as fixture from "@/storybook/fixture";
 
-import type { CardFormValues } from "../model/useCardForm";
-import { CardForm } from "./CardForm";
+import { CardFields, type CardFormFields } from "./CardFields";
 
-interface CardFormStoryProps {
+interface CardFieldsStoryProps {
   card: Card;
-  isSaving: boolean;
   validationError: boolean;
-  onCancel: () => void;
 }
 
-const CardFormStory = ({ card, isSaving, validationError, onCancel }: CardFormStoryProps) => {
-  const form = useForm<CardFormValues>({
+const CardFieldsStory = ({ card, validationError }: CardFieldsStoryProps) => {
+  const form = useForm<CardFormFields>({
     defaultValues: { frontText: card.frontText, backText: card.backText, tags: card.tags },
   });
 
@@ -28,42 +25,27 @@ const CardFormStory = ({ card, isSaving, validationError, onCancel }: CardFormSt
       form.setError("frontText", { message: "Front text is required." });
       form.setError("backText", { message: "Back text is required." });
     }
-    if (isSaving) void form.handleSubmit(() => new Promise(() => undefined))();
-  }, [form, isSaving, validationError]);
+  }, [form, validationError]);
 
-  return (
-    <CardForm
-      cardInfo={{
-        id: card.id,
-        uniqueKey: card.uniqueKey,
-        ...(card.createdAt ? { createdAt: card.createdAt } : {}),
-        ...(card.lastSeenAt != null ? { lastSeenAt: card.lastSeenAt } : {}),
-      }}
-      categories={CATEGORY}
-      form={form}
-      onCancel={onCancel}
-      onSubmit={form.handleSubmit(() => undefined)}
-    />
-  );
+  return <CardFields categories={CATEGORY} form={form} />;
 };
 
 const longCard = { ...fixture.card.long, tags: [...fixture.tags.toolong] };
 
 const meta = {
-  title: "Pages/Card Form/CardForm",
-  component: CardFormStory,
+  title: "Features/Card Form/CardFields",
+  component: CardFieldsStory,
   tags: ["autodocs"],
   parameters: { layout: "fullscreen" },
   decorators: [withPageLayout],
-  args: { card: fixture.card.default, isSaving: false, validationError: false, onCancel: fn() },
-} satisfies Meta<typeof CardFormStory>;
+  args: { card: fixture.card.default, validationError: false },
+} satisfies Meta<typeof CardFieldsStory>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {};
 export const ValidationError: Story = { args: { validationError: true } };
-export const Saving: Story = { args: { isSaving: true } };
 export const LongContent: Story = { args: { card: longCard } };
 export const Interaction: Story = {
   play: async ({ canvas, userEvent }) => {
