+++ src/features/card-form/ui/CardFields.tsx
@@ -2,10 +2,8 @@ import type * as React from "react";
 import { useId } from "react";
 import { type UseFormReturn, useFormState } from "react-hook-form";
 
-import type { CardId } from "@/entities/card";
-import { Button } from "@/shared/ui/button";
 import { TagList } from "@/shared/ui/content";
-import { Form, FormItem, Tag, Textarea } from "@/shared/ui/forms";
+import { FormItem, Tag, Textarea } from "@/shared/ui/forms";
 
 export interface CardFormFields {
   frontText: string;
@@ -13,24 +11,12 @@ export interface CardFormFields {
   tags: string[];
 }
 
-export interface CardFormProps {
-  cardInfo?: {
-    uniqueKey: string;
-    id: CardId;
-    createdAt?: number;
-    lastSeenAt?: number;
-  };
+export interface CardFieldsProps {
   categories: readonly string[];
   form: UseFormReturn<CardFormFields>;
-  onCancel: () => void;
-  onSubmit: React.SubmitEventHandler<HTMLFormElement>;
-  submitLabel?: string;
-  submittingLabel?: string;
 }
 
-const formatDate = (timestamp: number): string => new Date(timestamp).toLocaleDateString();
-
-export const CardForm: React.FC<CardFormProps> = (props) => {
+export const CardFields: React.FC<CardFieldsProps> = (props) => {
   const formState = useFormState({ control: props.form.control });
   const sectionHeadingIdPrefix = useId();
   const frontHeadingId = `${sectionHeadingIdPrefix}-card-front-heading`;
@@ -40,12 +26,9 @@ export const CardForm: React.FC<CardFormProps> = (props) => {
   const frontErrorId = `${frontInputId}-error`;
   const backInputId = `${sectionHeadingIdPrefix}-card-back-text`;
   const backErrorId = `${backInputId}-error`;
-  const submitLabel = props.submitLabel ?? "Save changes";
-  const submittingLabel = props.submittingLabel ?? "Saving…";
-  const submitContent = formState.isSubmitting ? <span>{submittingLabel}</span> : <span>{submitLabel}</span>;
 
   return (
-    <Form onSubmit={props.onSubmit}>
+    <>
       <section
         aria-labelledby={frontHeadingId}
         className="space-y-4 rounded-surface border border-border bg-surface p-4 md:p-5"
@@ -122,43 +105,6 @@ export const CardForm: React.FC<CardFormProps> = (props) => {
           ))}
         </TagList>
       </section>
-      {props.cardInfo !== undefined && (
-        <details className="rounded-surface border border-border bg-surface-muted p-4">
-          <summary className="flex min-h-touch cursor-pointer items-center font-semibold text-ink">
-            Card information
-          </summary>
-          <dl className="mt-4 grid gap-3 text-caption">
-            <div className="min-w-0">
-              <dt className="font-medium text-ink-muted">Unique key</dt>
-              <dd className="break-all text-ink">{props.cardInfo.uniqueKey}</dd>
-            </div>
-            <div className="min-w-0">
-              <dt className="font-medium text-ink-muted">ID</dt>
-              <dd className="break-all text-ink">{props.cardInfo.id}</dd>
-            </div>
-            {props.cardInfo.createdAt !== undefined && (
-              <div>
-                <dt className="font-medium text-ink-muted">Created</dt>
-                <dd className="text-ink">{formatDate(props.cardInfo.createdAt)}</dd>
-              </div>
-            )}
-            {props.cardInfo.lastSeenAt !== undefined && (
-              <div>
-                <dt className="font-medium text-ink-muted">Last seen</dt>
-                <dd className="text-ink">{formatDate(props.cardInfo.lastSeenAt)}</dd>
-              </div>
-            )}
-          </dl>
-        </details>
-      )}
-      <div className="flex flex-wrap justify-end gap-2 border-t border-border pt-4">
-        <Button variant="quiet" type="button" onClick={props.onCancel}>
-          Cancel
-        </Button>
-        <Button variant="primary" type="submit" disabled={formState.isSubmitting}>
-          {submitContent}
-        </Button>
-      </div>
-    </Form>
+    </>
   );
 };
