+++ src/pages/card-create/index.ts
@@ -1,0 +1,1 @@
+export { CardCreatePage } from "./ui/CardCreatePage";
