+++ src/pages/card-form/index.ts
@@ -1,2 +1,1 @@
-export { CardCreatePage } from "./ui/CardCreatePage";
 export { CardFormPage } from "./ui/CardFormPage";
